Refactor app with separate component for the Options component
